Remove the now redundant PackageSystemRequestDetails class.
