+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/packages/JpaPackageCache.java
@@ -34,6 +34,7 @@ import ca.uhn.fhir.jpa.model.entity.NpmPackageEntity;
 import ca.uhn.fhir.jpa.model.entity.NpmPackageVersionEntity;
 import ca.uhn.fhir.jpa.model.entity.NpmPackageVersionResourceEntity;
 import ca.uhn.fhir.jpa.model.entity.ResourceTable;
+import ca.uhn.fhir.jpa.partition.SystemRequestDetails;
 import ca.uhn.fhir.rest.api.Constants;
 import ca.uhn.fhir.rest.api.EncodingEnum;
 import ca.uhn.fhir.rest.api.server.storage.ResourcePersistentId;
@@ -326,7 +327,7 @@ public class JpaPackageCache extends BasePackageCacheManager implements IHapiPac
 	private ResourceTable createResourceBinary(IBaseBinary theResourceBinary) {
 
 		if (myPartitionSettings.isPartitioningEnabled()) {
-			PackageSystemRequestDetails myRequestDetails = new PackageSystemRequestDetails();
+			SystemRequestDetails myRequestDetails = new SystemRequestDetails();
 			return (ResourceTable) getBinaryDao().create(theResourceBinary, myRequestDetails).getEntity();
  		} else {
 			return (ResourceTable) getBinaryDao().create(theResourceBinary).getEntity();
@@ -637,7 +638,7 @@ public class JpaPackageCache extends BasePackageCacheManager implements IHapiPac
 	private void deleteAndExpungeResourceBinary(IIdType theResourceBinaryId, ExpungeOptions theOptions) {
 
 		if (myPartitionSettings.isPartitioningEnabled()) {
-			PackageSystemRequestDetails myRequestDetails = new PackageSystemRequestDetails();
+			SystemRequestDetails myRequestDetails = new SystemRequestDetails();
 			getBinaryDao().delete(theResourceBinaryId, myRequestDetails).getEntity();
 			getBinaryDao().forceExpungeInExistingTransaction(theResourceBinaryId, theOptions, myRequestDetails);
 		} else {
